init commit, code cleanup
