+++ tests/codeforces343d.rs
@@ -3,7 +3,7 @@
 //! module's contents directly here instead of the use statements.
 //! Also, use the commented code in main() to employ standard I/O.
 extern crate contest_algorithms;
-use contest_algorithms::graph::Graph;
+use contest_algorithms::graph::graph::Graph;
 use contest_algorithms::range_query::{specs::AssignSum, StaticArq};
 use contest_algorithms::scanner::Scanner;
 use std::io;
